requirements.txt
allure-pytest==2.13.1
pandas==2.0.0
pytest==7.2.0
selenium==4.7.2
webdriver-manager==3.8.5
